fix: search query for CharacterService
