+++ src/UI/loader/Loader.component.tsx
@@ -1,11 +1,7 @@
-import React from 'react';
-
 import './Loader.css';
 
-class Loader extends React.Component {
-  render() {
-    return <div className="loader"></div>;
-  }
-}
+const Loader = () => {
+  return <div className="loader"></div>;
+};
 
 export default Loader;
