+++ src/components/body/BodyView.component.tsx
@@ -1,0 +1,14 @@
+import { BodyProps } from '../../shared/interfaces';
+import Header from '../header/Header.component';
+import Main from '../main/Main.component';
+
+const BodyView = ({ updateCharacters, characters }: BodyProps) => {
+  return (
+    <>
+      <Header onSearch={updateCharacters} />
+      <Main characters={characters} />
+    </>
+  );
+};
+
+export default BodyView;
